Use template literals and cleaner size calculation
